Extracting START_TEST and ACTIVATE_TEST to tests module
