+++ src/components/test-slot/test-outcome/test-outcome.ts
@@ -2,7 +2,8 @@ import { Subscription } from 'rxjs/Subscription';
 import { Component, Input, OnInit } from '@angular/core';
 import { NavController, ModalController, Modal } from 'ionic-angular';
 import { Store, select } from '@ngrx/store';
-import { StartTest, ActivateTest, ResumingWriteUp } from '../../../pages/journal/journal.actions';
+import { StartTest, ActivateTest } from '../../../modules/tests/tests.actions';
+import { ResumingWriteUp } from '../../../pages/journal/journal.actions';
 import { TestStatus } from '../../../modules/tests/test-status/test-status.model';
 import { StartE2EPracticeTest } from '../../../pages/fake-journal/fake-journal.actions';
 import { startsWith } from 'lodash';
