+++ src/modules/tests/test-centre/test-centre.selector.ts
@@ -1,3 +1,12 @@
 import { TestCentre } from '@dvsa/mes-test-schema/categories/B';
+import { TestSlot } from '@dvsa/mes-journal-schema';
 
 export const getCostCentre = (testCentre: TestCentre) => testCentre.costCode || '';
+
+export const extractTestCentre = (slotData: TestSlot): TestCentre => {
+  return {
+    centreId: slotData.testCentre.centreId,
+    costCode: slotData.testCentre.costCode,
+    centreName: slotData.testCentre.centreName,
+  };
+};
