+++ src/modules/tests/test-data/__tests__/test-data.effects.spec.ts
@@ -3,7 +3,7 @@ import { ReplaySubject } from 'rxjs/ReplaySubject';
 import { TestBed } from '@angular/core/testing';
 import { provideMockActions } from '@ngrx/effects/testing';
 import * as testDataActions from '../test-data.actions';
-import * as journalActions from '../../../../pages/journal/journal.actions';
+import * as testsActions from '../../tests.actions';
 import { testsReducer } from '../../tests.reducer';
 import { Store, StoreModule } from '@ngrx/store';
 import { StoreModel } from '../../../../shared/models/store.model';
@@ -42,7 +42,7 @@ describe('Test Data Effects', () => {
       };
       const throttleAddDrivingFault = new testDataActions.ThrottleAddDrivingFault(faultPayload);
       // ARRANGE - setup the store
-      store$.dispatch(new journalActions.StartTest(123456));
+      store$.dispatch(new testsActions.StartTest(123456));
       store$.dispatch(throttleAddDrivingFault);
       // ACT - replay the action for the effect
       actions$.next(throttleAddDrivingFault);
@@ -58,7 +58,7 @@ describe('Test Data Effects', () => {
     it('should dispatch an action to toggle eco to be completed', (done) => {
       const toggleControlEcoAction = new testDataActions.ToggleControlEco();
       // ARRANGE - setup the store
-      store$.dispatch(new journalActions.StartTest(123456));
+      store$.dispatch(new testsActions.StartTest(123456));
       store$.dispatch(toggleControlEcoAction);
       // ACT - replay the action for the effect
       actions$.next(toggleControlEcoAction);
@@ -74,7 +74,7 @@ describe('Test Data Effects', () => {
     it('should dispatch an action to toggle eco to be completed', (done) => {
       const togglePlanningEcoAction = new testDataActions.TogglePlanningEco();
       // ARRANGE - setup the store
-      store$.dispatch(new journalActions.StartTest(123456));
+      store$.dispatch(new testsActions.StartTest(123456));
       store$.dispatch(togglePlanningEcoAction);
       // ACT - replay the action for the effect
       actions$.next(togglePlanningEcoAction);
