+++ src/modules/tests/test-slot-attributes/test-slot-attributes.selector.ts
@@ -7,6 +7,7 @@ export const isExtendedTest = (attributes: TestSlotAttributes) => attributes.ext
 export const isSpecialNeeds = (attributes: TestSlotAttributes) => attributes.specialNeeds || false;
 export const getSlotId = (attributes: TestSlotAttributes) => attributes.slotId;
 export const isWelshTest = (attributes: TestSlotAttributes) => attributes.welshTest;
+
 export const extractTestSlotAttributes = (slotData): TestSlotAttributes => ({
   welshTest: slotData.booking.application.welshTest,
   slotId: slotData.slotDetail.slotId,
