+++ src/providers/rekey-search/rekey-search.ts
@@ -1,26 +1,29 @@
 import { Injectable } from '@angular/core';
 import { RekeySearchParams } from './rekey-search.model';
-import { HttpClient } from '@angular/common/http';
-import { UrlProvider } from '../url/url';
+// import { HttpClient } from '@angular/common/http';
+// import { UrlProvider } from '../url/url';
 import { Observable } from 'rxjs/Observable';
+import { bookedTestMockHash } from './__mocks__/booked-test.mock';
+import { of } from 'rxjs/observable/of';
 
 @Injectable()
 export class RekeySearchProvider {
 
   constructor(
-    private httpClient: HttpClient,
-    private urlProvider: UrlProvider,
+    // private httpClient: HttpClient,
+    // private urlProvider: UrlProvider,
   ) { }
 
   getTest(rekeySearchParams: RekeySearchParams): Observable<Object> {
-    return this.httpClient.get(
-      this.urlProvider.getRekeySearchUrl(rekeySearchParams.staffNumber),
-      {
-        params: {
-          appRef: rekeySearchParams.applicationReference,
-        },
-      },
-    );
+    return of(bookedTestMockHash);
+    // return this.httpClient.get(
+    //   this.urlProvider.getRekeySearchUrl(rekeySearchParams.staffNumber),
+    //   {
+    //     params: {
+    //       appRef: rekeySearchParams.applicationReference,
+    //     },
+    //   },
+    // );
   }
 
 }
